+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ### Unreleased
 
+### [2.1.1] - 2026-06-02
+
+- refactor(rfc1869): defer address grammar to address parser
 - test fixes to mollify codeql
 
 ### [2.1.0] - 2026-05-27
@@ -92,3 +95,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 [1.1.4]: https://github.com/haraka/haraka-utils/releases/tag/v1.1.4
 [2.0.0]: https://github.com/haraka/haraka-utils/releases/tag/v2.0.0
 [2.1.0]: https://github.com/haraka/haraka-utils/releases/tag/v2.1.0
+[2.1.1]: https://github.com/haraka/haraka-utils/releases/tag/v2.1.1